remove BOM from input file
this one took me a sec. encodings are weird. i couldn't figure out why
the first field was being skipped until i jumped into a console and
checked the bytes. a little googling later and i found out that windows
adds a BOM to the beginning of files.

in a production system, i'd want to do a little more research to make
sure that there aren't other encoding issues we'd run across, but for
now, i'm just stripping them.