split out base transformer
normally, i prefer composition over inheritance, but for things like
this, i think there's an argument for this sort of pattern. it wouldn't
matter much for a one-off, but how many times is an import script like
this actually one and done?

in my experience, between different clients, different vendors, (even
different systems in-house), this sort of data mapping tends to be an
ongoing thing, so building a toolkit to handle things in a common way
can be a big win.