refactor to use DSL
okay, this is a big change, but i think it provides a much nicer dx.

the DSL is pretty simple, and i think it's pretty easy to read. you
can add a new field by calling `field` with the name of the field.
fields default to being treated as strings, but there are a few
different formatters available, such as `:date` and `:phone`.

another neat thing this allows is the ability to do custom and one-off
formatting by passing a block to a field. the type's formatter will run
first, and then the result will be passed to the block letting you
do whatever you want with it.

this also adds rudimentary validation. by default, fields are required,
but you can make them optional by passing `optional: true` to the field
definition. there is a _lot_ more room for improvement here, but i think
this illustrates the idea. in practice, i'd probably reach for active-
support for this.

i also added a `transform_batch` method to the base transformer. because
we record the order of fields as they're defined in the DSL, we can
transform batches of records in the same order. we add a :[] method to
access them by name, that way we don't have to worry about collisions
of field names with existing methods.