add basic reporter
i'm running out of time here, so this is pretty simple, but it works.

it leverages the batch transform method to keep track of the number of
transformed records and invalid records. it also keeps track of the
errors for each invalid record, as well as the total running time.

in a production system, i would probably use a logger instead of writing
to a file, but this is a quick and easy way to get the information out
of a given run.

the reporter is passed to the batch transform method, allowing you to
inject your own reporter if you want to do something different or do
some testing with it.

i think with more time, i'd also probably take in some arguments to
allow you control how verbose you want the report to be.