sketch out some more specs for the formatters
this definitely isn't exhaustive, but it shows the kinds of things
i'd want to best testing for with the formatters. i think i'd like
to extract the formatters into their own classes instead of just
having a bunch of lambdas. that would clean things up, but it would
also make it easier to test them in isolation.

alas, no time for that today!